Extract getNumberWidth from TOC/styled
