+++ src/app/content/components/TableOfContents/styled/index.tsx
@@ -4,26 +4,11 @@ import { CollapseIcon, ExpandIcon, iconSize } from '../../../../components/Detai
 import { labelStyle } from '../../../../components/Typography';
 import theme from '../../../../theme';
 import { ArchiveTree } from '../../../types';
-import { splitTitleParts } from '../../../utils/archiveTreeUtils';
 import ContentLinkComponent from '../../ContentLink';
 import { Tree } from 'react-aria-components';
+import getNumberWidth, { dividerWidth } from '../utils';
 
 export { ExpandIcon, CollapseIcon } from '../../../../components/Details.legacy';
-
-/* to regenerate these numbers, run this in a rex browser window
-(
-  (element) => 'wW1234567890.'.split('').reduce(
-    (result, char) => (result[char] = ((element.innerText = char) && element.getBoundingClientRect().width)) && result,
-    {}
-  )
-)(element = document.querySelector('[data-testid=toc] a[aria-current="page"]')
-  .appendChild(document.createElement('span'))
-);
- */
-const numberCharacterWidth = .890625;
-const letterCharacterWidth = 1.5109375;
-const numberPeriodWidth = .4453125;
-const dividerWidth = .8;
 
 const activeState = css`
   color: ${theme.color.text.black};
@@ -81,23 +66,6 @@ export const SummaryWrapper = styled.div`
     margin-top: 1.7rem;
   `)}
 `;
-
-const getNumberWidth = (contents: ArchiveTree['contents']) => contents.reduce((result, { title }) => {
-  const num = splitTitleParts(title)[0];
-
-  if (!num) {
-    return result;
-  }
-  const letters = num.replace(/[^A-Z]/ig, '');
-  const numbers = num.replace(/[^0-9]/g, '');
-  const periods = num.replace(/[^.]/g, '');
-
-  return Math.max(result,
-    numbers.length * numberCharacterWidth +
-    letters.length * letterCharacterWidth +
-    periods.length * numberPeriodWidth
-  );
-}, 0);
 
 export const StyledTree = styled(Tree)``;
